import dunctions

+++ database/migrations/2022_09_08_111751_create_price_codes_table.php
@@ -18,7 +18,7 @@ class CreatePriceCodesTable extends Migration
             $table->unsignedBigInteger('company_id')->nullable();
             $table->unsignedBigInteger('product_id')->nullable();
             $table->string('code');
-            $table->double('selling_price');
+            $table->decimal('selling_price', 10, 2);
             $table->string('price_basis');
             $table->timestamps();
